Refactored SoilServices. Created Extensions class. Started old to new api doc.
