+++ API/Models/BasicInfo.cs
@@ -1,0 +1,7 @@
+namespace API.Models;
+public class BasicInfo
+{
+    public string Name;
+    public double Latitude;
+    public double Longitude;
+}
